// apps/public/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      payments: {
        Row: {
          admin_notes: string | null
          amount: number
          booking_id: string | null
          cart_id: string | null
          client_id: string
          created_at: string
          currency: string
          id: string
          payment_date: string | null
          payment_method: string
          refund_amount: number | null
          refund_date: string | null
          status: string
          stripe_payment_intent_id: string | null
          stripe_session_id: string | null
          transaction_id: string | null
          updated_at: string
        }
        Insert: {
          admin_notes?: string | null
          amount: number
          booking_id?: string | null
          cart_id?: string | null
          client_id: string
          created_at?: string
          currency?: string
          id?: string
          payment_date?: string | null
          payment_method: string
          refund_amount?: number | null
          refund_date?: string | null
          status?: string
          stripe_payment_intent_id?: string | null
          stripe_session_id?: string | null
          transaction_id?: string | null
          updated_at?: string
        }
        Update: {
          admin_notes?: string | null
          amount?: number
          booking_id?: string | null
          cart_id?: string | null
          client_id?: string
          created_at?: string
          currency?: string
          id?: string
          payment_date?: string | null
          payment_method?: string
          refund_amount?: number | null
          refund_date?: string | null
          status?: string
          stripe_payment_intent_id?: string | null
          stripe_session_id?: string | null
          transaction_id?: string | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "payments_booking_id_fkey"
            columns: ["booking_id"]
            isOneToOne: false
            referencedRelation: "bookings"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "payments_booking_id_fkey"
            columns: ["booking_id"]
            isOneToOne: false
            referencedRelation: "missions_without_providers_in_zone"
            referencedColumns: ["booking_id"]
          },
          {
            foreignKeyName: "payments_cart_id_fkey"
            columns: ["cart_id"]
            isOneToOne: false
            referencedRelation: "carts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}